# cabmate-backend/openapi.yaml
openapi: 3.0.3
info:
  title: CabMate API
  description: |
    A campus carpooling API that lets students create, discover, and share rides.

    ## 🚀 Quick Start
    1. Expand **🧪 Try It Out** below → hit `GET /demo/credentials` → **Execute**
    2. Copy the `demo_user_id` and `ride_id` values from the response
    3. Paste them into any endpoint's **Try it out** fields — done!

    ## Endpoints
    - **Rides** – Create, search, view, delete rides
    - **Join Requests** – Request/approve/reject ride joins
    - **Profiles** – User profile CRUD
    - **Dashboard** – User-specific ride data

    ## Auth
    Pass `user_id` in request body or query params. Use `/demo/credentials` to get a test ID.

    ---
    Built by **Mohammad Alman Farooqui** | [Live App](https://cabmate.pages.dev)
  version: 1.0.0
  contact:
    name: Mohammad Alman Farooqui
    url: https://cabmate.pages.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://cabmate-auaw.onrender.com
    description: Production Server (Render)
  - url: http://localhost:8000
    description: Local Development Server
...
paths:
...
  /rides/:
    post:
      tags: [Rides]
      summary: Create a new ride
      description: |
        Creates a new ride listing. The ride will automatically expire 2 hours
        after the departure time via MongoDB TTL index.
      operationId: createRide
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideCreate"
            example:
              from_location: "Manipal University Jaipur"
              to_location: "Jaipur Railway Station"
              departure_time: "2026-03-15T10:30:00Z"
              seats_available: 3
              price_per_seat: 150
              remark: "Leaving sharp at 10:30, meet at Main Gate"
              created_by: "demo-swagger-test-user"
      responses:
        "200":
          description: Ride created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: MongoDB ObjectId of the created ride
                    example: "65a1b2c3d4e5f6a7b8c9d0e1"
        "500":
          description: Server error during ride creation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Rides]
      summary: Get all rides
      description: |
        Retrieves all rides, optionally filtered by location. Results are sorted
        by departure time (nearest first). Guest user rides are excluded.
      operationId: getRides
      parameters:
        - name: from_location
          in: query
          required: false
          schema:
            type: string
          description: Filter by exact origin location
          example: "Manipal University Jaipur"
        - name: to_location
          in: query
          required: false
          schema:
            type: string
          description: Filter by exact destination location
          example: "Jaipur Railway Station"
      responses:
        "200":
          description: List of rides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideWithCreator"
  /rides/search:
    get:
      tags: [Rides]
      summary: Search rides
      description: |
        Searches for upcoming rides with case-insensitive partial matching on locations.
        Only returns rides with departure time in the future. Guest user rides are excluded.
      operationId: searchRides
      parameters:
        - name: from_location
          in: query
          required: false
          schema:
            type: string
          description: Partial match for origin (case-insensitive)
          example: "Manipal"
        - name: to_location
          in: query
          required: false
          schema:
            type: string
          description: Partial match for destination (case-insensitive)
          example: "Railway"
      responses:
        "200":
          description: List of matching rides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideResponse"
  /rides/my-created:
    get:
      tags: [Rides]
      summary: Get rides created by a user
      description: Returns all rides created by the specified user.
      operationId: getMyCreatedRides
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: "User ID — use `demo-swagger-test-user` from /demo/credentials"
          example: "demo-swagger-test-user"
      responses:
        "200":
          description: List of rides created by the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideResponse"
  /rides/user/{user_id}:
    get:
      tags: [Rides]
      summary: Get user dashboard rides
      description: |
        Returns a combined dashboard view with:
        - **created**: All rides the user has created
        - **joined**: All rides where the user's join request was approved
      operationId: getUserDashboardRides
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: "User ID — use `demo-swagger-test-user` from /demo/credentials"
          example: "demo-swagger-test-user"
      responses:
        "200":
          description: Dashboard data with created and joined rides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
...
  /rides/{ride_id}:
    get:
      tags: [Rides]
      summary: Get a single ride
      description: Retrieves full details for a specific ride, including creator profile.
      operationId: getSingleRide
      parameters:
        - name: ride_id
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the ride
          example: "65a1b2c3d4e5f6a7b8c9d0e1"
      responses:
        "200":
          description: Ride details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideWithCreator"
        "400":
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Invalid ID format"
        "404":
          description: Ride not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Ride not found"

    delete:
      tags: [Rides]
      summary: Delete a ride
      description: |
        Deletes a ride and all associated join requests.
        Only the ride creator can delete their own ride (ownership check).
      operationId: deleteRide
      parameters:
        - name: ride_id
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the ride to delete
          example: "65a1b2c3d4e5f6a7b8c9d0e1"
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: "User ID of the ride creator — use `demo-swagger-test-user`"
          example: "demo-swagger-test-user"
      responses:
        "200":
          description: Ride deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ride deleted successfully"
        "400":
          description: Invalid Ride ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the ride creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "You can only delete your own rides"
        "404":
          description: Ride not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Join Requests ─────────────────────────────────────────────────────
  /rides/{ride_id}/request:
    post:
      tags: [Join Requests]
      summary: Request to join a ride
      description: |
        Sends a request to join a ride. The ride creator will then approve or reject it.
        - Cannot request to join your own ride
        - Cannot send duplicate requests
      operationId: requestToJoin
      parameters:
        - name: ride_id
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the ride
          example: "65a1b2c3d4e5f6a7b8c9d0e1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                  description: Supabase Auth UID of the requester
            example:
              user_id: "xyz789-requester-uid"
      responses:
        "200":
          description: Request sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Request sent"
        "400":
          description: |
            Bad request – possible reasons:
            - `user_id` not provided
            - Invalid ride ID
            - Creator cannot request their own ride
            - Already requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ride not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rides/{ride_id}/requests:
    get:
      tags: [Join Requests]
      summary: Get all requests for a ride
      description: |
        Returns all join requests for a specific ride, enriched with requester
        profile info (name, phone). Used by ride creators to manage requests.
      operationId: getRideRequests
      parameters:
        - name: ride_id
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the ride
          example: "65a1b2c3d4e5f6a7b8c9d0e1"
      responses:
        "200":
          description: List of join requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideRequest"
  /rides/{ride_id}/requests/{requester_id}/approve:
    post:
      tags: [Join Requests]
      summary: Approve a join request
      description: |
        Approves a pending join request. This also decrements the available seats
        by 1. Fails if no seats are available.
      operationId: approveRequest
      parameters:
        - name: ride_id
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the ride
          example: "65a1b2c3d4e5f6a7b8c9d0e1"
        - name: requester_id
          in: path
          required: true
          schema:
            type: string
          description: Supabase Auth UID of the requester to approve
          example: "xyz789-requester-uid"
      responses:
        "200":
          description: Request approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Request approved"
        "400":
          description: Invalid Ride ID or no seats available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ride or request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rides/{ride_id}/requests/{requester_id}/reject:
    post:
      tags: [Join Requests]
      summary: Reject a join request
      description: Rejects a pending join request. Seat count remains unchanged.
      operationId: rejectRequest
      parameters:
        - name: ride_id
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId of the ride
          example: "65a1b2c3d4e5f6a7b8c9d0e1"
        - name: requester_id
          in: path
          required: true
          schema:
            type: string
          description: Supabase Auth UID of the requester to reject
          example: "xyz789-requester-uid"
      responses:
        "200":
          description: Request rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Request rejected"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Profiles ──────────────────────────────────────────────────────────
...
components:
  schemas:
    RideCreate:
      type: object
      required:
        - from_location
        - to_location
        - departure_time
        - seats_available
        - price_per_seat
        - created_by
      properties:
        from_location:
          type: string
          description: Origin location / pickup point
          example: "Manipal University Jaipur"
        to_location:
          type: string
          description: Destination location
          example: "Jaipur Railway Station"
        departure_time:
          type: string
          format: date-time
          description: Scheduled departure time (ISO 8601 format, UTC)
          example: "2026-02-15T10:30:00Z"
        seats_available:
          type: integer
          minimum: 1
          description: Number of seats available for sharing
          example: 3
        price_per_seat:
          type: integer
          minimum: 0
          description: Price per seat in INR (₹)
          example: 150
        remark:
          type: string
          nullable: true
          description: Optional remark or instructions from the ride creator
          example: "Leaving sharp at 10:30, meet at Main Gate"
        created_by:
          type: string
          description: "User ID of the ride creator — use `demo-swagger-test-user` for testing"
          example: "demo-swagger-test-user"
    RideResponse:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId
          example: "65a1b2c3d4e5f6a7b8c9d0e1"
        from_location:
          type: string
          example: "Manipal University Jaipur"
        to_location:
          type: string
          example: "Jaipur Railway Station"
        departure_time:
          type: string
          format: date-time
          example: "2026-02-15T10:30:00+00:00"
        expires_at:
          type: string
          format: date-time
          description: Auto-expiry time (departure_time + 2 hours)
          example: "2026-02-15T12:30:00+00:00"
        seats_available:
          type: integer
          example: 3
        price_per_seat:
          type: integer
          example: 150
        remark:
          type: string
          nullable: true
          example: "Leaving sharp at 10:30, meet at Main Gate"
        created_by:
          type: string
          example: "abc123-user-uid"
        created_at:
          type: string
          format: date-time
          example: "2026-02-13T16:00:00+00:00"
    RideWithCreator:
      type: object
      description: Ride object with embedded creator profile (created_by field is replaced with creator object)
      properties:
        _id:
          type: string
          example: "65a1b2c3d4e5f6a7b8c9d0e1"
        from_location:
          type: string
          example: "Manipal University Jaipur"
        to_location:
          type: string
          example: "Jaipur Railway Station"
        departure_time:
          type: string
          format: date-time
          example: "2026-02-15T10:30:00+00:00"
        expires_at:
          type: string
          format: date-time
          example: "2026-02-15T12:30:00+00:00"
        seats_available:
          type: integer
          example: 3
        price_per_seat:
          type: integer
          example: 150
        remark:
          type: string
          nullable: true
          example: "Leaving sharp at 10:30"
        creator:
          $ref: "#/components/schemas/CreatorProfile"
        created_at:
          type: string
          format: date-time
          example: "2026-02-13T16:00:00+00:00"
    CreatorProfile:
      type: object
      nullable: true
      description: Embedded creator profile info
      properties:
        full_name:
          type: string
          example: "Alman Farooqui"
        phone:
          type: string
          example: "+91 9876543210"
        email:
          type: string
          example: "alman@example.com"
    DashboardResponse:
      type: object
      description: Dashboard data containing rides created and joined by the user
      properties:
        created:
          type: array
          description: Rides created by the user
          items:
            $ref: "#/components/schemas/RideWithCreator"
        joined:
          type: array
          description: Rides the user has joined (approved requests)
          items:
            $ref: "#/components/schemas/RideWithCreator"
    RideRequest:
      type: object
      description: A join request for a ride
      properties:
        _id:
          type: string
          description: MongoDB ObjectId of the request
          example: "65b2c3d4e5f6a7b8c9d0e1f2"
        ride_id:
          type: string
          description: ID of the ride being requested
          example: "65a1b2c3d4e5f6a7b8c9d0e1"
        requester_id:
          type: string
          description: Supabase Auth UID of the requester
          example: "xyz789-requester-uid"
        status:
          type: string
          enum: [pending, approved, rejected]
          description: Current status of the request
          example: "pending"
        requested_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-02-13T17:00:00+00:00"
        requester:
          type: object
          nullable: true
          description: Requester profile info (name and phone)
          properties:
            full_name:
              type: string
              example: "John Doe"
            phone:
              type: string
              example: "+91 9123456789"
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: "Ride not found"